requirements.txt
#nsml: mlvclab/pytorch:1.5-cuda10.1-cudnn7-devel
adamp==0.2.0
anyconfig==0.9.11
asn1crypto==1.3.0
backcall==0.1.0
beautifulsoup4==4.9.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
conda==4.8.3
conda-build==3.18.11
conda-package-handling==1.6.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
docopt==0.6.2
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
gitdb==4.0.5
GitPython==3.1.3
glob2==0.7
greenlet==0.4.16
idna==2.8
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.7.0
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonpickle==1.4.1
kiwisolver==1.2.0
libarchive-c==2.9
lief==0.9.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
munch==2.5.0
networkx==2.4
nose==1.3.7
numpy==1.18.1
olefile==0.46
opencv-python==4.3.0.36
packaging==20.4
pandas==1.0.5
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pkginfo==1.5.0.1
portalocker==1.7.1
prompt-toolkit==3.0.4
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
py-cpuinfo==6.0.0
pyclipper==1.2.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.6.1
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.22.0
ruamel-yaml==0.15.87
sacred==0.8.1
scikit-image==0.17.2
scipy==1.5.0
Shapely==1.7.0
six==1.14.0
smmap==3.0.4
soupsieve==2.0
tensorboardX==2.1
tifffile==2020.7.17
torch==1.5.0
torch-encoding==1.2.1
torchvision==0.6.0a0+82fd1c8
tqdm==4.42.1
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
